add option to add markers to genotype protocol using VCF upload

+++ lib/SGN/Controller/AJAX/GenotypesVCFUpload.pm
@@ -86,9 +86,6 @@ sub upload_genotype_verify_POST : Args(0) {
 
     my $project_id = $c->req->param('upload_genotype_project_id') || undef;
     my $protocol_id = $c->req->param('upload_genotype_protocol_id') || undef;
-#    print STDERR "PROJECT ID =".Dumper($project_id)."\n";
-#    print STDERR "PROTOCOL ID =".Dumper($protocol_id)."\n";
-
     my $organism_species = $c->req->param('upload_genotypes_species_name_input');
     my $protocol_description = $c->req->param('upload_genotypes_protocol_description_input');
     my $project_name = $c->req->param('upload_genotype_vcf_project_name');
@@ -103,10 +100,15 @@ sub upload_genotype_verify_POST : Args(0) {
     my $reference_genome_name = $c->req->param('upload_genotype_vcf_reference_genome_name');
     my $assay_type = $c->req->param('assay_type_select');
     my $add_new_accessions = $c->req->param('upload_genotype_add_new_accessions');
+    my $add_new_markers = $c->req->param('upload_genotype_add_new_markers');
     my $add_accessions;
     if ($add_new_accessions){
         $add_accessions = 1;
         $obs_type = 'accession';
+    }
+    my $add_markers;
+    if ($add_new_markers) {
+	$add_markers = 1;
     }
     my $include_igd_numbers;
     if ($contains_igd){
@@ -184,10 +186,12 @@ sub upload_genotype_verify_POST : Args(0) {
     my $timestamp = $time->ymd()."_".$time->hms();
 
     my $upload_original_name;
+    my $transposed_basename;
     my $upload_tempfile;
     my $subdirectory;
     my $parser_plugin;
     if ($upload_vcf) {
+
         $upload_original_name = $upload_vcf->filename();
         $upload_tempfile = $upload_vcf->tempname;
         $subdirectory = "genotype_vcf_upload";
@@ -227,9 +231,11 @@ sub upload_genotype_verify_POST : Args(0) {
             close($F);
             close($Fout);
             $upload_tempfile = $temp_file_transposed;
-            $upload_original_name = basename($temp_file_transposed);
+            $transposed_basename = basename($temp_file_transposed);
             $parser_plugin = 'transposedVCF';
-        }
+        } else {
+	    $transposed_basename = basename($upload_vcf->tempname);
+	}
     }
     if ($upload_transposed_vcf) {
         $upload_original_name = $upload_transposed_vcf->filename();
@@ -246,7 +252,7 @@ sub upload_genotype_verify_POST : Args(0) {
             tempfile => $upload_tempfile,
             subdirectory => $subdirectory,
             archive_path => $c->config->{archive_path},
-            archive_filename => $upload_original_name,
+            archive_filename => $transposed_basename,
             timestamp => $timestamp,
             user_id => $user_id,
             user_role => $user_role
@@ -297,7 +303,7 @@ sub upload_genotype_verify_POST : Args(0) {
         close($F);
         close($Fout);
         $upload_tempfile = $temp_file_transposed;
-        $upload_original_name = basename($temp_file_transposed);
+        $transposed_basename = basename($temp_file_transposed);
 
         $subdirectory = "genotype_transposed_vcf_upload";
         $parser_plugin = 'transposedVCF';
@@ -311,7 +317,7 @@ sub upload_genotype_verify_POST : Args(0) {
         $parser_plugin = 'IntertekCSV';
 
         if ($obs_type eq 'accession') {
-            $include_lab_numbers = 1;
+	    $include_lab_numbers = 0;
         }
 
         my $upload_inteterk_marker_info_original_name = $upload_inteterk_marker_info->filename();
@@ -373,24 +379,30 @@ sub upload_genotype_verify_POST : Args(0) {
         unlink $upload_kasp_marker_info_tempfile;
     }
 
-    my $uploader = CXGN::UploadFile->new({
-        tempfile => $upload_tempfile,
-        subdirectory => $subdirectory,
-        archive_path => $c->config->{archive_path},
-        archive_filename => $upload_original_name,
-        timestamp => $timestamp,
-        user_id => $user_id,
-        user_role => $user_role
-    });
-    my $archived_filename_with_path = $uploader->archive();
-    my $md5 = $uploader->get_md5($archived_filename_with_path);
-    if (!$archived_filename_with_path) {
-        push @error_status, "Could not save file $upload_original_name in archive.";
-        return (\@success_status, \@error_status);
+    my $archived_filename_with_path;
+    my $md5;
+    if ($upload_vcf) {
+        my $uploader = CXGN::UploadFile->new({
+            tempfile => $upload_tempfile,
+            subdirectory => $subdirectory,
+            archive_path => $c->config->{archive_path},
+            archive_filename => $transposed_basename,
+            timestamp => $timestamp,
+            user_id => $user_id,
+            user_role => $user_role
+        });
+        $archived_filename_with_path = $uploader->archive();
+        $md5 = $uploader->get_md5($archived_filename_with_path);
+        if (!$archived_filename_with_path) {
+            push @error_status, "Could not save file $upload_original_name in archive.";
+            return (\@success_status, \@error_status);
+        } else {
+            push @success_status, "File $upload_original_name saved in archive.";
+        }
+        unlink $upload_tempfile;
     } else {
-        push @success_status, "File $upload_original_name saved in archive.";
+	print STDERR "archive_path not set\n";
     }
-    unlink $upload_tempfile;
 
     #if protocol_id provided, a new one will not be created
     if ($protocol_id){
@@ -401,7 +413,7 @@ sub upload_genotype_verify_POST : Args(0) {
         $organism_species = $protocol->species_name;
         $obs_type = $protocol->sample_observation_unit_type_name;
         if ($obs_type eq 'accession') {
-            $include_lab_numbers = 1;
+	    $include_lab_numbers = 0;
         }
     }
 
@@ -429,8 +441,9 @@ sub upload_genotype_verify_POST : Args(0) {
         observation_unit_type_name => $obs_type,
         organism_id => $organism_id,
         create_missing_observation_units_as_accessions => $add_accessions,
+	create_missing_markers => $add_markers,
         igd_numbers_included => $include_igd_numbers,
-        # lab_numbers_included => $include_lab_numbers
+        lab_numbers_included => $include_lab_numbers
     });
     $parser->load_plugin($parser_plugin);
 
@@ -538,48 +551,63 @@ sub upload_genotype_verify_POST : Args(0) {
 
             if ($protocol_id) {
                 my @protocol_match_errors;
+		my @mismatch_markers;
+
                 my $new_marker_data = $protocol->{markers};
+
                 my $stored_protocol = CXGN::Genotype::Protocol->new({
                     bcs_schema => $schema,
                     nd_protocol_id => $protocol_id
                 });
-                my $stored_markers = $stored_protocol->markers();
 
-                my @all_stored_markers = keys %$stored_markers;
-                my %compare_marker_names = map {$_ => 1} @all_stored_markers;
-                my @mismatch_marker_names;
-                while (my ($chrom, $new_marker_data_1) = each %$new_marker_data) {
-                    while (my ($marker_name, $new_marker_details) = each %$new_marker_data_1) {
-                        if (exists($compare_marker_names{$marker_name})) {
-                            while (my ($key, $value) = each %$new_marker_details) {
-                                if ($value ne ($stored_markers->{$marker_name}->{$key})) {
-                                    push @protocol_match_errors, "Marker $marker_name in your file has $value for $key, but in the previously stored protocol shows ".$stored_markers->{$marker_name}->{$key};
-                                }
+                my $stored_markers = $stored_protocol->markers();
+                my %stored_marker_names = map { $_ => 1 } keys %{$stored_markers};
+
+		for my $chrom (keys %{$new_marker_data}) {
+                    my $markers_on_chrom = $new_marker_data->{$chrom};
+
+		    for my $marker_name (keys %{$markers_on_chrom}) {
+                        my $new_marker_details = $markers_on_chrom->{$marker_name};
+
+			unless ($stored_marker_names{$marker_name}) {
+                            push @mismatch_markers, [$chrom, $marker_name];
+                            next;
+                        }
+
+			my $stored_details = $stored_markers->{$marker_name};
+
+                        for my $key (keys %{$new_marker_details}) {
+                            my $new_value    = defined $new_marker_details->{$key}
+                                ? $new_marker_details->{$key}
+                                : '';
+                            my $stored_value = defined $stored_details->{$key}
+                                ? $stored_details->{$key}
+                                : '';
+
+                            if ($new_value ne $stored_value) {
+                                push @protocol_match_errors,
+                                    "Marker $marker_name in your file has $new_value for $key, "
+                                  . "but in the previously stored protocol shows $stored_value";
                             }
-                        } else {
-                            push @mismatch_marker_names, $marker_name;
                         }
                     }
                 }
 
-                if (scalar(@mismatch_marker_names) > 0){
-                    my $marker_name_error;
-                    $marker_name_error .= "<br>";
-                    foreach my $error ( sort @mismatch_marker_names) {
-                        $marker_name_error .= $error."<br>";
-                    }
-
-                    $c->stash->{rest} = { error => "These marker names in your file are not in the selected protocol. $marker_name_error"};
-                    $c->detach();
+                if (@mismatch_markers) {
+		    if ($add_markers) {
+			print STDERR "Adding new markers\n";
+			    $store_genotypes->store_new_markers_in_protocolprop(\@mismatch_markers);
+		    } else {
+			my $marker_name_error = join '<br>', map { $_->[0] . ': ' . $_->[1] } @mismatch_markers;
+                        $c->stash->{rest} = { error => "These marker names in your file are not in the selected protocol. $marker_name_error"};
+                        $c->detach();
+		    }
                 }
 
-
-                if (scalar(@protocol_match_errors) > 0){
-                    my $protocol_warning;
-                    foreach my $match_error (@protocol_match_errors) {
-                        $protocol_warning .= $match_error."<br>";
-                    }
-                    if (!$accept_warnings){
+                if (@protocol_match_errors) {
+                    my $protocol_warning = join '<br>', @protocol_match_errors;
+                    
+	            if (!$accept_warnings){
                         $c->stash->{rest} = { warning => $protocol_warning };
                         $c->detach();
                     }
